Consider mazes with loops
The previous solution made the assumption thare was at most one path to reach a
valid point.
This did not hold for some tests, and this is updated.

Another solution did not involve looping twice in free_path:
I thought of in the already_seen queue enqueueing a struct holding both
a point_t struct and a flag saying whether this point was on the path or not.
However, I did not think I could do that with our linter.